resource: update to align with common KBS design
