+++ src/resources/mod.rs
@@ -1,3 +1,5 @@
 pub mod directory_key_manager;
 pub mod gpg;
 pub mod image;
+pub mod file;
+pub mod opa;
